ENH: rework much of the electron display:
- Main benefit is now savefig should be very reliable
- I changed the behavior so that immediately after displaying a plot, I
ask plotly.js to do async operations to generate the svg for the figure.
I then store this in a global variable (yuck!) and just access it when I
want to savefig
- The downside here is that any restyle! or relayout! operations won't
be included in the svg data. I think I can probably overcome that, but
It isn't totally obvious to me how yet...